+++ internal/grpcx/invoke_cover_test.go
@@ -1,3 +1,5 @@
+//go:build grpclive
+
 package grpcx
 
 import (